Import tools from external file
